[issue #1147] Renamed cameraview to camViz
