+++ src/tools/camViz/camViz.glade
@@ -4,7 +4,7 @@
 <glade-interface>
   <widget class="GtkWindow" id="mainwindow">
     <property name="visible">True</property>
-    <property name="title" translatable="yes">Cameraview</property>
+    <property name="title" translatable="yes">CamViz</property>
     <child>
       <widget class="GtkTable" id="table1">
         <property name="visible">True</property>
